// src/utils/file-type/Types.d.ts
export type BufferEncoding = "ascii" | "utf-8" | "binary";